アセット以外のゲームジャムに出した状態のものです。

+++ Assets/Tanimura/BulletScript.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class BulletScript : MonoBehaviour
 {
-    public static int _speed = 2000;
+    public static float _speed = 2000;
     public void Shoot(Vector3 dir)
     {
         GetComponent<Rigidbody>().AddForce(dir);
